dieu chinh quyen sua phieu - theo phong ban

+++ CCS/BUS/t_account.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using DAL;
 using System.Data.Linq;
+using System.Windows.Forms;
 
 namespace BUS
 {
@@ -15,7 +16,7 @@ namespace BUS
         {
             db = new KetNoiDBDataContext();
             var dn = (from tk in db.accounts where tk.uname == taikhoan & tk.pass == matkhau select tk).ToList();
-                return dn;
+            return dn;
         }
         public List<account> dangnhap2(string taikhoan, string matkhau)
         {
@@ -24,7 +25,7 @@ namespace BUS
             return dn;
         }
 
-        public void moi (string id, string uname, string name, string pass, string phongban, string madonvi, bool ia, bool khopxm)
+        public void moi(string id, string uname, string name, string pass, string phongban, string madonvi, bool ia, bool khopxm)
         {
             account ac = new account();
             ac.id = id;
@@ -41,7 +42,7 @@ namespace BUS
         public void sua(string id, string uname, string name, string pass, string phongban, string madonvi, bool ia, bool khopxm)
         {
             account ac = (from tb in db.accounts select tb).Single(t => t.id == id);
-       
+
             ac.uname = uname;
             ac.name = name;
             ac.pass = pass;
@@ -49,15 +50,22 @@ namespace BUS
             ac.madonvi = madonvi;
             ac.khopxm = khopxm;
             ac.IsActived = ia;
-   
+
             db.SubmitChanges();
         }
 
         public void xoa(string id)
         {
             account ac = (from tb in db.accounts select tb).Single(t => t.id == id);
-            db.accounts.DeleteOnSubmit(ac);
-            db.SubmitChanges();
-        } 
+            if (ac.id == "AD" || ac.id == "ADMIN")
+            {
+                MessageBox.Show("BẠN KHÔNG THỂ XOÁ TÀI KHOẢN ADMIN !");
+            }
+            else
+            {
+                db.accounts.DeleteOnSubmit(ac);
+                db.SubmitChanges();
+            }
+        }
     }
 }
